+++ primefaces/src/main/java/org/primefaces/model/badge/DefaultBadgeModel.java
@@ -36,6 +36,7 @@ public class DefaultBadgeModel implements BadgeModel, Serializable {
     private String styleClass;
     private String icon;
     private String iconPos;
+    private String onClick;
     private boolean visible = true;
 
     @Override
@@ -118,6 +119,16 @@ public class DefaultBadgeModel implements BadgeModel, Serializable {
         this.iconPos = iconPos;
     }
 
+    @Override
+    public String getOnclick() {
+        return onClick;
+    }
+
+    @Override
+    public void setOnclick(String onClick) {
+        this.onClick = onClick;
+    }
+
     public static Builder builder() {
         return new Builder();
     }
@@ -165,6 +176,11 @@ public class DefaultBadgeModel implements BadgeModel, Serializable {
             return this;
         }
 
+        public Builder onClick(String onClick) {
+            badgeModel.setOnclick(onClick);
+            return this;
+        }
+
         public Builder visible(boolean visible) {
             badgeModel.setVisible(visible);
             return this;
@@ -174,5 +190,4 @@ public class DefaultBadgeModel implements BadgeModel, Serializable {
             return badgeModel;
         }
     }
-
 }